fix: inputs
