+++ rs-week1/src/components/pages/formPage/FormInput.tsx
@@ -1,86 +1,27 @@
 import './FormPage.css';
-import { IFormInputProps, IFormInputState } from 'components/types/interface';
-import React, { Component, RefObject } from 'react';
-import { validatePrice, validateDate, validateImageFile, validateProductStatus } from './Validate';
-import { Ecategory } from '../../types/enum';
+import { IFormInputProps, FormData } from 'components/types/interface';
+import React, { useState } from 'react';
+import { validateDate } from './Validate';
 import InputCategory from './InputComponents/InputCategory';
 import InputProductStatus from './InputComponents/InputProductStatus';
 import InputImage from './InputComponents/InputImage';
 import { useForm } from 'react-hook-form';
 
-interface FormData {
-  title: string;
-  price: number;
-  description: string;
-  date: string;
-  rules: boolean;
-  productStatus: string;
-  imageFile: FileList;
-  category: string;
-}
-
 function FormInput({ onChangeProduct }: IFormInputProps) {
+  const [statusValid, setStatusValid] = useState(false);
+
   const {
     register,
     handleSubmit,
     reset,
+    watch,
     formState: { errors },
   } = useForm<FormData>({ mode: 'onSubmit', reValidateMode: 'onSubmit' });
-  //  const checkValidAllInputs = () => {
-  //     const priceValid = validatePrice(this.inputPriceRef, /^\d+$/);
-  //     const titleValid = validateText(this.inputTitleRef);
-  //     const descriptionValid = validateText(this.inputDescriptionRef);
-  //     const dateValid = validateDate(this.inputDateRef);
-  //     const imageFileValid = validateImageFile(this.inputFileRef);
-  //     const categoryValid = !!Object.values(Ecategory).find(
-  //       (cat) => cat === this.inputCategoryRef.current?.value
-  //     );
-  //     const productStatusValid = validateProductStatus(
-  //       this.inputProductStatusRefNew,
-  //       this.inputProductStatusRefUsed
-  //     );
-  //     const rulesValid = this.inputRulesRef.current?.checked ?? false;
-  //     this.setState({ spanPriceValid: priceValid });
-  //     this.setState({ spanTitleValid: titleValid });
-  //     this.setState({ spanDescriptionValid: descriptionValid });
-  //     this.setState({ spanDateValid: dateValid });
-  //     this.setState({ spanFileValid: imageFileValid });
-  //     this.setState({ spancategoryValid: categoryValid });
-  //     this.setState({ spanProductStatusValid: productStatusValid });
-  //     this.setState({ spanRulesValid: rulesValid });
-  //     if (
-  //       priceValid &&
-  //       titleValid &&
-  //       descriptionValid &&
-  //       dateValid &&
-  //       imageFileValid &&
-  //       categoryValid &&
-  //       productStatusValid &&
-  //       rulesValid
-  //     ) {
-  //       this.setState({ statusValid: true });
-  //       return true;
-  //     }
-  //     return false;
-  //   };
 
-  // getImageInput = () => {
-  //   let url = '';
-  //   if (this.inputFileRef.current && this.inputFileRef) {
-  //     const imageFile = this.inputFileRef.current.files?.[0];
-  //     if (imageFile) {
-  //       url = URL.createObjectURL(imageFile);
-  //       this.setState({
-  //         imageUrl: url,
-  //         imageFile,
-  //       });
-  //     }
-  //   }
-  //   return url;
-  // };
+  const fileName = watch('imageFile')?.[0]?.name || 'no name of image';
 
   const onSubmit = (data: FormData) => {
-    console.log('data', data);
+    setStatusValid(true);
     onChangeProduct({
       id: Date.now(),
       title: data.title,
@@ -88,73 +29,15 @@ function FormInput({ onChangeProduct }: IFormInputProps) {
       date: data.date,
       productStatus: data.productStatus,
       description: data.description,
-      imageUrl: data.title,
+      imageUrl: URL.createObjectURL(data.imageFile[0]),
       category: data.category,
     });
     reset();
-    // event.preventDefault();
-    // const imageUrl = this.getImageInput();
-    // const validForm = this.checkValidAllInputs();
-
-    // if (!validForm) return;
-
-    // const newTitle = this.inputTitleRef.current?.value ?? '';
-    // const newDescription = this.inputDescriptionRef.current?.value ?? '';
-    // const newPrice = this.inputPriceRef.current?.value ?? '';
-    // const newDate = this.inputDateRef.current?.value ?? '';
-    // const newSelect = this.inputProductStatusRefNew.current?.checked;
-    // const usedSelect = this.inputProductStatusRefUsed.current?.checked;
-
-    // let newProductStatus = '';
-
-    // if (newSelect && !usedSelect && this.inputProductStatusRefNew.current) {
-    //   newProductStatus = this.inputProductStatusRefNew.current?.value;
-    // } else if (!newSelect && usedSelect && this.inputProductStatusRefUsed.current) {
-    //   newProductStatus = this.inputProductStatusRefUsed.current?.value;
-    // }
-
-    // const newCategory = this.inputCategoryRef.current?.value ?? '';
-
-    // const newProduct = {
-    //   id: Date.now(),
-    //   title: newTitle,
-    //   price: Number(newPrice),
-    //   date: newDate,
-    //   productStatus: newProductStatus,
-    //   description: newDescription,
-    //   imageUrl,
-    //   category: newCategory,
-    // };
-
-    // this.props.onChangeProduct(newProduct);
-
-    // this.setDefaultForm();
-
-    // setTimeout(() => {
-    //   this.setState({ statusValid: false });
-    // }, 5000);
+    setTimeout(() => {
+      setStatusValid(false);
+    }, 5000);
   };
 
-  // setDefaultForm = () => {
-  //   if (this.formRef) this.formRef.current?.reset();
-  //   this.setState({
-  //     imageUrl: '',
-  //     imageFile: null,
-  //   });
-  // };
-
-  // const {
-  //   imageFile,
-  //   spanFileValid,
-  //   spanPriceValid,
-  //   spanTitleValid,
-  //   spancategoryValid,
-  //   spanDescriptionValid,
-  //   spanDateValid,
-  //   spanProductStatusValid,
-  //   spanRulesValid,
-  //   statusValid,
-  // } = this.state;
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
       <div className="form-input">
@@ -213,7 +96,11 @@ function FormInput({ onChangeProduct }: IFormInputProps) {
           placeholder="description product: Phone is very..."
         />
       </div>
-      {/* <InputImage valid={spanFileValid} forwardRef={this.inputFileRef} imageFile={imageFile} /> */}
+      <InputImage
+        error={errors.imageFile}
+        register={register('imageFile', { required: 'true' })}
+        fileName={fileName}
+      />
       <div className="form-input">
         <label htmlFor="date-input">
           Date of manufacture:
@@ -253,7 +140,7 @@ function FormInput({ onChangeProduct }: IFormInputProps) {
       <InputCategory error={errors.category} register={register('category', { required: true })} />
       <div className="form-input">
         <label htmlFor="category-select">
-          {/* Status: {statusValid && <span style={{ color: 'green' }}>Card added</span>} */}
+          Status: {statusValid && <span style={{ color: 'green' }}>Card added</span>}
         </label>
         <button type="submit">Submit CARD</button>
       </div>
@@ -262,8 +149,3 @@ function FormInput({ onChangeProduct }: IFormInputProps) {
 }
 
 export default FormInput;
-
-// validate: {
-//   acceptedFormat: (files: FileList | null) =>
-//     (files && ['image/jpg', 'image/png'].includes(files[0].type)) || 'error',
-// },
